Bidding functionality, model FloatField updated to DecimalField

+++ auction/forms.py
@@ -2,12 +2,10 @@ from django import forms
 from django.db.models import Max
 
 
-
-
 class CreateAuctionForm(forms.Form):
     title = forms.CharField(max_length=255)
     description = forms.CharField(widget=forms.Textarea())
-    minimum_price = forms.FloatField(min_value=0.01)
+    minimum_price = forms.DecimalField(decimal_places=2, max_digits=15, min_value=0.01)   # allow up to 1 000 000 000 000.00
     deadline_date = forms.DateTimeField(
         input_formats=['%d.%m.%Y %H:%M:%S'],
         widget=forms.TextInput(attrs={"placeholder": "dd.mm.yyyy HH:MM:SS"}),
